+++ generateme/colors.clj
@@ -2,7 +2,11 @@
   (:require [clojure2d.core :refer :all]
             [clojure2d.color :as c]
             [fastmath.core :as m]
-            [fastmath.vector :as v]))
+            [fastmath.vector :as v]
+            [clojure2d.extra.utils :as u]
+            [fastmath.clustering :as cl]
+            [clojure2d.pixels :as p]
+            [fastmath.stats :as stat]))
 
 (def canv (canvas 800 300))
 
@@ -18,3 +22,17 @@
         (line c x 0 x 150)
         (set-color c (p1 t))
         (line c x 150 x 300)))))
+
+;;
+
+(def img (p/load-pixels "docs/samurai.jpg"))
+
+(u/show-palette (c/reduce-colors :LUV (resize (p/load-pixels "bbb1.jpg") 600 600) 5) )  
+
+;;
+
+(save (:buffer (u/show-gradient (c/gradient :LAB :cubic-spline (c/reduce-colors :LAB (resize (p/load-pixels "b5.jpg") 600 600) 5)))) "b5grad.jpg")
+
+(u/show-gradient (c/gradient :LAB :cubic-spline (c/reduce-colors :LUV (resize (p/load-pixels "bbb1.jpg") 600 600) 4)))
+
+;;